+++ requirements_checker/__init__.py
@@ -1,3 +1,3 @@
 from .check_string import *
-from .check_requirements_txt import *
+from .check_pkg import *
 from .check_version import *